module handler / session type pairing complete
handler annotations + message label is used to find the correct session type branch that corresponds to a handler fn

no parsing/transformation of session type strings to structs - for development I'm just manually building a big lookup table and indexing the table using the handler label to get all possible branches. The branches are then checked one at a time to find which label matches the branch and information to uniquely id a handler is returned, otherwise the first error is returned.